added file to make sure that if ROL is not installed everything still works
